### 2021-12-17 #4001 @whedon
[PRE REVIEW]: Octopus-sensing: A python library for human behavior studies
**Submitting author:** @nastaran62 (<a href="http://orcid.org/0000-0002-6082-9772">Nastaran Saffaryazdi</a>)
**Repository:** <a href="https://github.com/octopus-sensing/octopus-sensing" target ="_blank">https://github.com/octopus-sensing/octopus-sensing</a>
**Version:** v3.1.0
