### 2021-12-17 #4001 @kyleniemeyer
Hi @nastaran62, thanks for your interest in JOSS. Unfortunately, we do not have any editors from this area available at the moment, so this submission will be on our waitlist until someone becomes available to edit.

In the meantime, could you please take a look at the missing DOIs in your article and resolve those? Also, any reviewer suggestions you could make would be helpful. Thanks!
