+++ paper/paper.bib
@@ -2,6 +2,7 @@
   title={Autonomic nervous system activity in emotion: A review},
   author={Kreibig, Sylvia D},
   journal={Biological psychology},
+  doi={10.1016/j.biopsycho.2010.03.010},
   volume={84},
   number={3},
   pages={394--421},
@@ -12,6 +13,7 @@
 @article{chen2020physiological,
   title={Physiological linkage during shared positive and shared negative emotion.},
   author={Chen, Kuan-Hua and Brown, Casey L and Wells, Jenna L and Rothwell, Emily S and Otero, Marcela C and Levenson, Robert W and Fredrickson, Barbara L},
+  doi={10.1037/pspi0000337},
   journal={Journal of Personality and Social Psychology},
   year={2020},
   publisher={American Psychological Association}
@@ -21,6 +23,7 @@
   title={Multimodal Affective State Assessment Using fNIRS+ EEG and Spontaneous Facial Expression},
   author={Sun, Yanjia and Ayaz, Hasan and Akansu, Ali N},
   journal={Brain Sciences},
+  doi={10.3390/brainsci10020085},
   volume={10},
   number={2},
   pages={85},
@@ -32,6 +35,7 @@
   title={Development of a Real-Time Emotion Recognition System Using Facial Expressions and EEG based on machine learning and deep neural network methods},
   author={Hassouneh, Aya and Mutawa, AM and Murugappan, M},
   journal={Informatics in Medicine Unlocked},
+  doi={10.1016/j.imu.2020.100372},
   pages={100372},
   year={2020},
   publisher={Elsevier}
@@ -41,6 +45,7 @@
   title={Psychopathy and physiological detection of concealed information: A review},
   author={Verschuere, Bruno and Crombez, Geert and Koster, Ernst and Uzieblo, Katarzyna},
   journal={Psychologica Belgica},
+  doi={10.5334/pb-46-1-2-99},
   volume={46},
   number={1-2},
   year={2006},
@@ -51,6 +56,7 @@
   title={Observers’ physiological measures in response to videos can be used to detect genuine smiles},
   author={Hossain, Md Zakir and Gedeon, Tom},
   journal={International Journal of Human-Computer Studies},
+  doi={10.1016/j.ijhcs.2018.10.003},
   volume={122},
   pages={232--241},
   year={2019},
@@ -61,6 +67,7 @@
   title={Engagement detection in online learning: a review},
   author={Dewan, M Ali Akber and Murshed, Mahbub and Lin, Fuhua},
   journal={Smart Learning Environments},
+  doi={10.1186/s40561-018-0080-z},
   volume={6},
   number={1},
   pages={1},
@@ -72,6 +79,7 @@
   title={Affective Robot Story-Telling Human-Robot Interaction: Exploratory Real-Time Emotion Estimation Analysis Using Facial Expressions and Physiological Signals},
   author={Val-Calvo, Mikel and {\'A}lvarez-S{\'a}nchez, Jos{\'e} Ram{\'o}n and Ferr{\'a}ndez-Vicente, Jos{\'e} Manuel and Fern{\'a}ndez, Eduardo},
   journal={IEEE Access},
+  doi={10.1109/ACCESS.2020.3007109},
   volume={8},
   pages={134051--134066},
   year={2020},
@@ -82,6 +90,7 @@
   title={A Multimodal Emotional Human-Robot Interaction Architecture for Social Robots Engaged in Bidirectional Communication$\backslash$vspace* 7pt},
[…]
