### 2022-01-10 #4045 @whedon
[REVIEW]: Octopus-sensing: A Python library for human behavior studies
**Submitting author:** @nastaran62 (<a href="http://orcid.org/0000-0002-6082-9772">Nastaran Saffaryazdi</a>)
**Repository:** https://github.com/octopus-sensing/octopus-sensing
**Branch with paper.md** (empty if default branch): master
**Version:** v3.1.1
