### 2022-01-28 #4045 @peircej
I'm able to tick "yes" to most of the requirements already. 🥳 A couple of things remain

**Community guidelines:** at the moment I can't see any information about how users/contributors should get involved with the project. There is the GitHub issues list, which has a single closed issue. There's a page in the docs for [developement](https://octopus-sensing.nastaran-saffar.me/development.html) but doesn't give any information about how to contribute - just information about how to pip install. I'd recommend the authors provide a line of information in the readme.md and also [CONTRIBUTING file with guidelines](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)

**State of the Field:** I can't see anything in the manuscript that considers other related packages. There are many packages around BCI using Python. There's Lab Streaming Layer for combining and time-stamping data from devices. There's PsychoPy for stimulus delivery. Maybe the authors should mention these and their relationships and/or their reasons for not integrating/extending those existing packages.
