### 2022-02-06 #4045 @ixjlyons
Overall, Octopus Sensing appears to be a nice library for getting up and running with experiments requiring simultaneous recording from multiple sensors. I came across a few (mostly minor) issues described below, but otherwise I recommend accepting.

### Minor issues

These are a few issues I noticed that might be good to have fixed prior to publication.

#### Documentation

- Typo: README, link text "Otopus Sensing Website"
- Broken link: Quick Start doc page link to "Development Guide"
- `sudo` for running `init.sh` doesn't seem necessary (Quick Start doc) and should be avoided if possible.
- Typo: docs tutorial section 3, first paragraph: "overal" -> "overall", "Uniti" -> "Unity", "taht" -> "that"
- Need another troubleshooting step for PyGObject? May want to refer to [their docs](https://pygobject.readthedocs.io/en/latest/getting_started.html). I had to pre-install a couple system packages to get `pygobject` to install successfully.
- Docs tutorial, section 9 header is ill-formatted.
- Broken link: Octopus Sensing Visualizer page, note: reference to preprocessing module is broken
- Stimuli API doc page is empty.

#### Code

- `init_script.sh` points to `nastaran62/octopus-sensing`

#### Paper, JOSS checklist

- Only first two paper authors are in the commit history. Maybe you could describe author contributions here in the review issue - I'll leave that to editor discretion.
- Last sentence of second paragraph: no space after "Octopus Sensing".
- Last sentence of third paragraph: could use grammar rework - unclear what "They" refers to (researchers or sensors?), particularly with phrases "making intelligent devices" and "making interactive robots".
- I think the paper and documentation might benefit from a brief description of the methods used to provide simultaneous, synchronous recording from multiple sensors (i.e. process-based device interface). This could help convey the level of timing precision and accuracy to be expected.
- Community guidelines appear to be missing.

### Other Comments

These are some general comments on the software, documentation, etc. They're intended to be suggestions for further improvements but I don't consider them necessary for publication.

- Lots of dependencies pulled in for devices I may or may not have. You might consider using something like `extras_require` for extra devices (not sure how that would work with Poetry).
- In early parts of the tutorial, you might want to include imports to demonstrate how the package is organized.
- Overall I think it could be helpful for the examples to run out-of-the-box if at all possible, whether it's by using only "standard" devices like camera streaming and audio that are expected to work without special hardware, or simulated devices just for demonstration. Being able to run an example without modification is an important step in evaluating software. As it is now, I don't think I can run any of the examples without at least some modifications. The `endpoint_example` only required setting up a directory with images - perhaps a small example image stack could be included in the repository.
- In the tutorials, diagrams showing how things are connected would be helpful to understand the various network entities involved.
- Docs tutorial, section 5: a link when referring to API for questionnaire and windows would be good. In general, links to API documentation are very helpful in tutorials/examples.
