### 2022-02-10 69abd73
Fixing devices bug, when because of network delay receive several start before receiving stop
