### 2022-02-11 0720a07
Apply changes in documents based on reviewers' comments
