+++ paper/paper.bib
@@ -119,18 +119,18 @@
   publisher={IEEE}
 }
 
-@article{Saffaryazdi2021,
-  title={Investigation of the use of Facial Micro-expressions in Combination with EEG and Physiological signals for Emotion Recognition},
-  author={Nastaran Saffaryazdi, Syed Talal Wasim, Suranga Nanayakkara and Mark Billinghurst},
+@article{Saffaryazdi2022,
+  title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
+  author={Nastaran Saffaryazdi, Syed Talal Wasim,  Kuldeep Dileep, Alireza Farrokhinia, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
   deartment={Auckland Bioengineering Intitute, University of Auckland},
-  publisher={Manuscript is under submission}
+  publisher={Manuscript is under review}
 }
 
-@article{Saffaryazdi2021conv,
-  title={Emotion Recognition in Conversations using Brain and Physiological Signals},
+@article{Saffaryazdi2022conv,
+  title={Emotion Recognition in Conversations Using Brain and Physiological Signals},
   author={Nastaran Saffaryazdi, Yenushka Goonesekera, Nafiseh Saffaryazdi and Nebiyou Daniel Hailemariam and Ebasa Girma Temesgen, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
   deartment={Auckland Bioengineering Intitute, University of Auckland},
-  publisher={Manuscript submitted for publication}
+  publisher={Manuscript is under publication}
 }
 
 @article{egger2019emotion,
