+++ paper/paper.md
@@ -36,23 +36,23 @@ bibliography: paper.bib
 ---
  
 # Summary
-Designing user studies and collecting data is critical in exploring and automatically recognizing human behavior. Currently, it is possible to use a wide range of sensors to capture heart rate, brain activity, skin conductance, and a variety of different physiological cues [@seneviratne2017survey]. These can be combined together to provide information about a user’s emotional state [@egger2019emotion; @dzedzickis2020human], cognitive load [@vanneste2020towards; @mangaroska2021exploring], or other factors. However, even when data are collected correctly, synchronizing data from multiple sensors is time-consuming and subject to human error. Failure to record and synchronize data can lead to incorrect analysis and results, and finally, the need to repeat the time-consuming experiments several times. 
+Designing user studies and collecting data is critical in exploring and automatically recognizing human behavior. Currently, it is possible to use a wide range of sensors to capture heart rate, brain activity, skin conductance, and a variety of different physiological cues [@seneviratne2017survey]. These can be combined to provide information about a user's emotional state [@egger2019emotion; @dzedzickis2020human], cognitive load [@vanneste2020towards; @mangaroska2021exploring], or other factors. However, even when data are collected correctly, synchronizing data from multiple sensors is time-consuming and subject to human error. Failure to record and synchronize data can lead to incorrect analysis and results, and finally, the need to repeat the time-consuming experiments several times. 
  
-To overcome these challenges, `Octopus Sensing` facilitates synchronous data acquisition from various sources and provides some utilities for designing user studies, real-time monitoring, and offline data visualization.  The major aim of `Octopus Sensing`is to provide a simple scripting interface so that people with little or no software development skills can define sensor-based experiment scenarios with minimum effort.
+To overcome these challenges, `Octopus Sensing` facilitates synchronous data acquisition from various sources and provides some utilities for designing user studies, real-time monitoring, and offline data visualization. The primary aim of `Octopus Sensing` is to provide a simple scripting interface so that people with little or no software development skills can define sensor-based experiment scenarios with minimum effort.
  
 # Statement of need
-External events affect the human body and mind, creating many internal and external changes in response to external stimuli. Nowadays, researchers use various sensors to monitor and measure these responses to know more about a person’s state [@kreibig2010autonomic; @chen2020physiological; @sun2020multimodal] and employ sensors in many healthcare applications like assisting patients or mental health monitoring [@hassouneh2020development]. They can also be used to improve social interactions [@verschuere2006psychopathy; @hossain2019observers], and creating better quality of life by making intelligent devices like Intelligent Tutoring Systems [@dewan2019engagement] or making interactive robots and virtual characters [@val2020affective; @hong2020multimodal].
+Several changes occur in the body and mind due to various internal and external stimuli. Nowadays, researchers use various sensors to measure and monitor these responses to determine an individual's state [@kreibig2010autonomic; @chen2020physiological; @sun2020multimodal] and to help patients [@hassouneh2020development] or monitor mental health [@africa2020psychiatry]. Monitoring and analyzing human responses can be used to improve social interactions [@verschuere2006psychopathy; @hossain2019observers] and improve quality of life by creating intelligent devices like Intelligent Tutoring Systems [@dewan2019engagement], creating adaptive systems [@aranha2019adapting], or creating interactive robots and virtual characters [@val2020affective; @hong2020multimodal].
  
[…]
