+++ docs/tutorial.rst
@@ -8,7 +8,7 @@ Tutorial
 What Are We Building?
 ----------------------
 
-In this tutorial, we'll show how to design a simple scenario with octopus-sensing step by step.
+This tutorial will show how to design a simple scenario with octopus-sensing step by step.
 
 The example scenario is the most common in emotion recognition research in affective computing. In this scenario, we learn how to record data from different sources synchronously when an event happens and stop data recording by finishing the event.
 
@@ -27,8 +27,8 @@ The example scenario is the most common in emotion recognition research in affec
 
 **Prerequisites**
 
-Create a project and install `octopus-sensing` package by following the instructions on :ref:`quick_start`. We recommend using `pipenv` to do so.
-And then copy the source of examples from `examples` package in octopus-sensing repository to your project directory and run them.
+Create a project and install the `octopus-sensing` package by following the instructions on :ref:`quick_start`. We recommend using `pipenv` to do so.
+And then, copy the source of examples from `examples` package in octopus-sensing repository to your project directory and run them.
 
 1- Record data from various sources synchronously
 -------------------------------------------------
@@ -49,7 +49,7 @@ Firstly we should create a Shimmer3Streaming object with a specific name and an 
 ...                                saving_mode=SavingModeEnum.CONTINIOUS_SAVING_MODE,
 ...                                output_path="./output")
 
-Then we should add the created object to the `DeviceCoordinator`.  As the name suggests, the device coordinator is responsible for coordination, like starting to record data in all devices at once, stopping data recording, triggering (marking data at a point), and terminating devices. When a device is added to the device coordinator, it will be initialized and prepared for recording.
+Then we should add the created object to the `DeviceCoordinator`. As the name suggests, the device coordinator is responsible for coordination, like starting to record data in all devices at once, stopping data recording, triggering (marking data at a point), and terminating devices. When a device is added to the device coordinator, it will be initialized and prepared for recording.
 
 >>> device_coordinator = DeviceCoordinator()
 >>> device_coordinator.add_devices([my_shimmer])
@@ -81,10 +81,10 @@ The name of the recorded file will be `shimmer-{experiment_id}.csv` and will be 
 Keep in your mind, before running the code, turn on the Shimmer3 sensor and pair Bluetooth and the serial port.
 (Shimmer password: 1234)
 
-For example, in Linux you can do it as follow:
+For example, in Linux, you can do it as follow:
     1. hcitool scan   //It shows the mac-address of the device. for shimmer it is 00:06:66:F0:95:95
     2. vim /etc/bluetooth/rfcomm.conf write the below line in it: rfcomm0{ bind no; device 00:06:66:F0:95:95; channel 1; comment "serial port" }
-    3. sudo rfcomm connect rfcomm0 00:06:66:F0:95:95 // This is for reading bluetooth data from a serial port
+    3. sudo rfcomm connect rfcomm0 00:06:66:F0:95:95 // This is for reading Bluetooth data from a serial port
 
 Adding several sensors
 """"""""""""""""""""""
@@ -205,28 +205,28 @@ See this example in **octopus-sensing/examples/remote_device_example/matlabRecor
 
 3- Receiving Messages over Network
 -----------------------------------
-Octopus Sensing provides an endpoint which by starting it, it listens for incoming Message requests.
+Octopus Sensing provides an endpoint that listens for incoming Message requests by starting it.
 It passes the message to the Device Coordinator to dispatch them to the devices.
 It accepts HTTP POST requests. The Body can be serialized in one of 'json', 'msgpack'
 or 'pickle'.
-This feature can be used when we have designed the overal scenario with other programming languages, or scenario
-is running in other software like Uniti or Matlab. In this cases, we should write a simple code in python taht uses
[…]
