+++ docs/api/octopus_sensing.stimuli.rst
@@ -5,3 +5,20 @@ Stimuli
    :members:
    :undoc-members:
    :show-inheritance:
+
+Image Stimulus
+--------------------------------------------
+
+.. automodule:: octopus_sensing.stimuli.image_stimulus
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Video Stimulus
+----------------------------------------------
+
+.. automodule:: octopus_sensing.stimuli.video_stimulus
+   :members:
+   :undoc-members:
+   :show-inheritance:

[…]
