### 2022-02-12 1ee9de9
Improve the init script. No 'sudo' required anymore.
