+++ init_script/main.py
@@ -1,3 +1,0 @@
-import octopus_sensing
-
-# Under construction...
