+++ docs/api/octopus_sensing.stimuli.rst
@@ -1,3 +1,5 @@
+.. _stimuli:
+
 Stimuli
 ================================
 
