+++ docs/tutorial.rst
@@ -40,19 +40,22 @@ In this section, we learn how to record data from different sensors with interna
 
 Adding a sensor
 """"""""""""""""
-Imagine you want to record physiological data using shimmer3 sensor by pressing a key on the keyboard
+Imagine you want to record video using your built-in webcam by pressing a key on the keyboard
 and stop recording after 5 seconds.
 
-Firstly we should create a Shimmer3Streaming object with a specific name and an output path for recording data.
+Firstly we should create a CameraStreaming object with a specific name and an output path for recording data.
 
->>> my_shimmer = Shimmer3Streaming(name="shimmer",
-...                                saving_mode=SavingModeEnum.CONTINIOUS_SAVING_MODE,
-...                                output_path="./output")
+>>> from octopus_sensing.devices import CameraStreaming
+>>> from octopus_sensing.device_coordinator import DeviceCoordinator
+>>> from octopus_sensing.common.message_creators import start_message, stop_message
+>>> my_camera = CameraStreaming(camera_no=0,
+...                             name="camera",
+...                             output_path="./output")
 
 Then we should add the created object to the `DeviceCoordinator`. As the name suggests, the device coordinator is responsible for coordination, like starting to record data in all devices at once, stopping data recording, triggering (marking data at a point), and terminating devices. When a device is added to the device coordinator, it will be initialized and prepared for recording.
 
 >>> device_coordinator = DeviceCoordinator()
->>> device_coordinator.add_devices([my_shimmer])
+>>> device_coordinator.add_devices([my_camera])
 
 We are now developing a simple code to start data recording by pressing a key and stopping recording after 5 seconds.
 
@@ -60,6 +63,7 @@ We are now developing a simple code to start data recording by pressing a key an
 >>> device_coordinator.dispatch(start_message(experiment_id, stimuli_id))
 >>> time.sleep(5)
 >>> device_coordinator.dispatch(stop_message(experiment_id, stimuli_id))
+>>> time.sleep(0.5)
 >>> device_coordinator.terminate()
 
 Octopus-sensing provides a set of default messages for handling different actions like
@@ -70,21 +74,8 @@ For example, we use the id of the recorded subject as experiment ID.
 Defining stimulus ID is essential for identifying the recorded data related to each stimulus
 when we have different stimuli.
 
-To see the completed example see **octopus_sensing/examples/add_sensors.py**.
-By running this example, according to the `saving_mode` option that we passed when creating the sensor instance,
-the recorded file/s will be different. The default value of saving mode for Shimmer3 is continuous.
-It means if we have several stimuli, all data will be recorded in one file.
-The name of the recorded file will be `shimmer-{experiment_id}.csv` and will be saved in `output/shimmer` path. In this file, Shimmer3 data samples have been recorded from when it initialized to when it received the terminate message. The last column of data is the trigger column, which shows in what sample and time the device has received the start and stop triggers (pressing the button and 5 seconds after that). If we change the saving mode to separate (`SavingModeEnum.SEPARATED_SAVING_MODE`), it will record one file for each stimulus (For this example, one file), and the name of stimuli will appear in the file name.
-
-**Troubleshooting**
-
-Keep in your mind, before running the code, turn on the Shimmer3 sensor and pair Bluetooth and the serial port.
-(Shimmer password: 1234)
-
-For example, in Linux, you can do it as follow:
-    1. hcitool scan   //It shows the mac-address of the device. for shimmer it is 00:06:66:F0:95:95
[…]
