+++ paper/paper.bib
@@ -187,5 +187,16 @@
   publisher={IEEE}
 }
 
+@article{aranha2019adapting,
+  title={Adapting software with affective computing: a systematic review},
+  author={Aranha, Renan Vinicius and Corr{\^e}a, Cl{\'e}ber Gimenez and Nunes, F{\'a}tima LS},
+  journal={IEEE Transactions on Affective Computing},
+  volume={12},
+  number={4},
+  pages={883--899},
+  year={2019},
+  publisher={IEEE}
+}
+
 
 
