+++ mkdocs.yml
@@ -1,5 +1,5 @@
 site_name: Octopus Sensing
-repo_url: https://github.com/nastaran62/octopus-sensing/
+repo_url: https://github.com/octopus-sensing/octopus-sensing
 edit_uri: ""
 site_author: "Nastaran Saffaryazdi"
 site_description: ""
