### 2022-02-14 ead8366
Add 'Adding Support for a New Device' section to development.rst
