### 2022-02-14 c8473f3
add test_camera_streaming.py in ./test
