+++ paper/paper.md
@@ -44,9 +44,9 @@ Several changes occur in the body and mind due to various internal and external 
  
 Researchers have recently attempted to gain a deeper understanding of humans by simultaneously studying physiological and behavioral changes in the human body [@shu2018review; @koelstra2011deap]. Acquiring and analyzing data from different sources with various hardware and software is complex, time-consuming, and challenging. Additionally, synchronously recording data in multiple formats can be easily affected by human error. These tasks slow down the pace of progress in human-computer interaction and human behavior research.  
  
-There are only a few frameworks that support synchronous data acquisition and design. [iMotions](https://imotions.com/) has developed software for integrating and synchronizing data recording through a wide range of various sensors and devices. Despite having many great features, iMotions is commercial software and not open-source. In contrast, there are a few open-source programs for conducting human studies. Psychopy (https://www.psychopy.org/) is a powerful open-source, cross-platform software that is mainly used for building experiments' pipelines in behavioral science with visual and auditory stimuli. It can also record data and send the triggers to a few devices that record human responses. Another effort in this area is LabStreamingLayer (LSL) (http://labstreaminglayer.org/), which provides synchronized, multimodal streaming. While a wide range of devices supports streaming data through LSL, an application still needs to be run for acquiring data from each sensor separately. The recording and synchronization of data are also affected by network delays and failures.
+There are only a few frameworks that support synchronous data acquisition and design. [iMotions](https://imotions.com/) has developed software for integrating and synchronizing data recording through a wide range of various sensors and devices. Despite having many great features, iMotions is commercial software and not open-source. In contrast, there are a few open-source programs for conducting human studies. [Psychopy](https://www.psychopy.org/) is a powerful open-source, cross-platform software that is mainly used for building experiments' pipelines in behavioral science with visual and auditory stimuli. It can also record data and send the triggers to a few devices software that record human responses. Another effort in this area is [LabStreamingLayer (LSL) LabRecorder](http://labstreaminglayer.org/). Although, LSL LabRecorder provides synchronized, multimodal streaming through a wide range of devices, an extra application still needs to be run for acquiring data from each sensor separately.
 
-We designed `Octopus Sensing` to minimize the effect of network failure in synchronous data streaming and reduce the number of applications that we should run for data streaming through different devices. `Octopus Sensing` is a lightweight open-source multi-platform library that facilitates synchronous data acquisition from various sources through a unified software and could be easily extended to process and analyze data in real-time. It provides a web-based real-time monitoring system that can be used remotely to illustrate and monitor signals in real-time. It also offers offline data visualization to see various human responses simultaneously.
+`Octopus Sensing` is a lightweight open-source multi-platform library that facilitates synchronous data acquisition from various sources through a unified software and could be easily extended to process and analyze data in real-time. We designed the 'Octopus Sensing` library to minimize the effect of network failure in synchronous data streaming and reduce the number of applications that we should run for data streaming through different devices. Rather than creating a standalone software or framework, we created a library that could be easily integrated with other applications. Octopus Sensing provides a real-time monitoring system for illustrating and monitoring signals remotely using a web-based platform. The system also provides offline data visualization to see various human responses at the same time.
  
 # Overview
  
@@ -61,17 +61,18 @@ The main features of `Octopus Sensing` are listed as follows:
 * Provides offline visualization of data from multiple sources simultaneously.
 
 # Methodology
[…]
