+++ README.md
@@ -35,7 +35,7 @@ Main features
 
 Copyright
 ---------
-Copyright © 2020 Nastaran Saffaryazdi, Aidin Gharibnavaz
+Copyright © 2020-2022 Nastaran Saffaryazdi, Aidin Gharibnavaz
 
 This program is free software: you can redistribute it and/or modify it under the terms of the GNU
 General Public License as published by the Free Software Foundation, either version 3 of the
