### 2022-02-16 69944fd
Fixing the overview image and paper writing
