+++ paper/paper.md
@@ -61,14 +61,14 @@ The main features of `Octopus Sensing` are listed as follows:
 * Provides offline visualization of data from multiple sources simultaneously.
 
 # Methodology
-Octopus Sensing synchronizes data recording by using `multiprocessing` in Python. It creates a message queue and a process with three threads for every device by creating the device's instance. The main thread handles the message queue. One thread is responsible for data streaming and reacording, and the other thread listens on a queue and when received a request, returns the last three second of data from a buffer. This data can be used in real-time monitoring, or real-time processing and creating real-time feedbacks.  The `Device Coordinator` sends different triggers such as the start of recording or end of recording to different devices by putting the message in all devices' queue. The `Device Coordinator` also sends the trigger over the network for devices that are not embedded in the Octopus Sensing. 
+Octopus Sensing synchronizes data recording by using `multiprocessing` in Python. By initiating an instance of a device, Octopus Sensing creates a process for the device. Each device's process has three threads: data acquisition thread, trigger handling and data recording thread, and real-time data thread. The data acquisition thread is responsible for acquiring data from a sensor. The trigger handling and data recording thread handles trigger messages through a message queue for synchronous data recording. It also records data in a file or files. The real-time data thread listens on a queue and returns the last three seconds of data from a buffer when receiving a request. This data is being used in real-time monitoring and can be used for real-time processing and creating real-time feedback in the future. The `Device Coordinator` sends different triggers such as the start of recording or end of recording to different devices by putting the message in all devices' trigger queues. The `Device Coordinator` also sends the trigger over the network for devices that are not embedded in the Octopus Sensing. The following image shows the overall view of the main components of the Octopus Sensing and their relations.
 <p align="center">
   <img src="https://github.com/octopus-sensing/octopus-sensing/blob/paper-review/docs/_static/OCS-diagram.png" alt="Ovrall view of Octopus Sensing">
 </p>
  
 # Research perspective
  
-We have used `Octopus Sensing` to design several human emotion recognition experiments. We designed the experiments and recorded facial video, brain activity, and physiological signals using `Octopus Sensing` in a watching video task to recognize emotion [@Saffaryazdi2021]. This scenario which is common in physiological emotion recognition studies has been included in the repository as an example and explained in the tutorial. In another research study, we collected multimodal data in a face-to-face conversation task to analyze human emotional responses [@Saffaryazdi2021conv].
+We used `Octopus Sensing` to design several human emotion recognition experiments. We designed the experiments and recorded facial video, brain activity, and physiological signals using `Octopus Sensing` in a watching video task to recognize emotion [@Saffaryazdi2021]. This scenario which is common in physiological emotion recognition studies has been included in the repository as an example and explained in the tutorial. In another research study, we collected multimodal data in a face-to-face conversation task to analyze human emotional responses [@Saffaryazdi2021conv].
  
 This tool can be used to build real-time data processing systems to recognize emotions, stress, cognitive load, or analyze human behavior. Our final goal is to extend its capabilities to provide real-time emotion recognition using multimodal data. Furthermore, we plan to integrate it with `Psychopy` in the future and combine multimodal data collection and monitoring with `Psychopy` features when designing scenarios. Additionally, we plan to support LSL in the future. By supporting LSL, other applications that already support LSL could work with Octopus Sensing.
  
