+++ paper/paper.bib
@@ -129,7 +129,8 @@
 @article{Saffaryazdi2022conv,
   title={Emotion Recognition in Conversations Using Brain and Physiological Signals},
   author={Nastaran Saffaryazdi, Yenushka Goonesekera, Nafiseh Saffaryazdi and Nebiyou Daniel Hailemariam and Ebasa Girma Temesgen, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
-  deartment={Auckland Bioengineering Intitute, University of Auckland},
+  doi={10.1145/3490099.3511148},
+  year={2022},
   publisher={Manuscript is under publication}
 }
 
