+++ paper/paper.bib
@@ -122,15 +122,15 @@
 @article{Saffaryazdi2022,
   title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
   author={Nastaran Saffaryazdi, Syed Talal Wasim,  Kuldeep Dileep, Alireza Farrokhinia, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
-  deartment={Auckland Bioengineering Intitute, University of Auckland},
+  deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
   publisher={Manuscript is under review}
 }
 
 @article{Saffaryazdi2022conv,
   title={Emotion Recognition in Conversations Using Brain and Physiological Signals},
   author={Nastaran Saffaryazdi, Yenushka Goonesekera, Nafiseh Saffaryazdi and Nebiyou Daniel Hailemariam and Ebasa Girma Temesgen, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
-  doi={10.1145/3490099.3511148},
   year={2022},
+  deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
   publisher={Manuscript is under publication}
 }
 
@@ -192,6 +192,7 @@
   title={Adapting software with affective computing: a systematic review},
   author={Aranha, Renan Vinicius and Corr{\^e}a, Cl{\'e}ber Gimenez and Nunes, F{\'a}tima LS},
   journal={IEEE Transactions on Affective Computing},
+  doi={10.1109/TAFFC.2019.2902379},
   volume={12},
   number={4},
   pages={883--899},
