### 2022-02-16 b4d73ab
Add image to the paper path and change the document
