### 2022-02-16 acbb82f
A quick review of the paper
