### 2022-02-16 a703033
Clarification about Windows in quick_start
