+++ docs/quick_start.rst
@@ -36,8 +36,8 @@ The created `main.py` file is an example application. To run it:
 
 If you don't want to use the script, you can use the following methods instead.
 
-Installation using Pipenv
-=========================
+Installation using Pipenv (All Platforms)
+=========================================
 
 We recommend using a package manager like `Pipenv <https://pipenv.pypa.io/>`_ instead of globally
 installing Octopus Sensing using `pip` to prevent package conflicts. To do so, follow these
@@ -62,8 +62,8 @@ It installs Octopus Sensing inside the virtual environment created by Pipenv. Yo
 
 Refer to `Pipenv website <https://pipenv.pypa.io/>`_ for more info.
 
-Installation using pip
-======================
+Installation using pip (All Platforms)
+======================================
 
 You can use `pip` to install `octopus-sensing` as simple as:
 
@@ -76,8 +76,8 @@ Then it can be imported like:
 :code:`import octopus_sensing`
 
 
-Installation from source
-========================
+Installation from source (All Platforms)
+========================================
 
 If you want to compile it from source for development purposes or to have the un-released features,
 please refer to :ref:`development`.
@@ -91,4 +91,4 @@ Troubleshooting
 You need to install `portaudio` package on your system. On a debian-based linux the package called
 `portaudio19-dev`.
 
-- Pip cannot install PyGObject on Windows. Users need to install it manually themselves. See `PyGObject documentation <https://pygobject.readthedocs.io/en/latest/getting_started.html#windows-getting-started>`_ to know how to install PyGObject on Windows.
+- Pip cannot install PyGObject on Windows. If users want to use `octopus-sensing.stimuli` or `octopus-sensing.windows` packages, they need to install it manually themselves. See `PyGObject documentation <https://pygobject.readthedocs.io/en/latest/getting_started.html#windows-getting-started>`_ to know how to install PyGObject on Windows.
