### 2022-02-16 34962e4
Remove portaudio.h from Troubleshooting because we no longer use it
