+++ paper/paper.bib
@@ -122,16 +122,19 @@
 @article{Saffaryazdi2022,
   title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
   author={Nastaran Saffaryazdi, Syed Talal Wasim,  Kuldeep Dileep, Alireza Farrokhinia, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
+  year={2022}
   deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
   publisher={Manuscript is under review}
+  pubstate = {Under review},
 }
 
 @article{Saffaryazdi2022conv,
   title={Emotion Recognition in Conversations Using Brain and Physiological Signals},
   author={Nastaran Saffaryazdi, Yenushka Goonesekera, Nafiseh Saffaryazdi and Nebiyou Daniel Hailemariam and Ebasa Girma Temesgen, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
   year={2022},
-  deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
-  publisher={Manuscript is under publication}
+  doi={10.1145/3490099.3511148},
+  publisher={27th International Conference on Intelligent User Interfaces (IUI '22), March 22--25, 2022, Helsinki, Finland}
+  pubstate = {Under publication},
 }
 
 @article{egger2019emotion,
