+++ paper/paper.bib
@@ -122,7 +122,7 @@
 @article{Saffaryazdi2022,
   title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
   author={Nastaran Saffaryazdi, Syed Talal Wasim,  Kuldeep Dileep, Alireza Farrokhinia, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
-  year={2022}
+  year={2022},
   deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
   publisher={Manuscript is under review}
   pubstate = {Under review},
