+++ paper/paper.bib
@@ -121,7 +121,7 @@
 
 @article{Saffaryazdi2022,
   title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
-  author={Nastaran Saffaryazdi, Syed Talal Wasim,  Kuldeep Dileep, Alireza Farrokhinia, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
+  author={Saffaryazdi, Wasim and Dileep, Farrokhinia and Nanayakkara, Broadbent and Billinghurst},
   year={2022},
   deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
   publisher={Manuscript is under review},
@@ -130,7 +130,7 @@
 
 @article{Saffaryazdi2022conv,
   title={Emotion Recognition in Conversations Using Brain and Physiological Signals},
-  author={Nastaran Saffaryazdi, Yenushka Goonesekera, Nafiseh Saffaryazdi and Nebiyou Daniel Hailemariam and Ebasa Girma Temesgen, Suranga Nanayakkara and Elizabeth Broadbent and Mark Billinghurst},
+  author={Saffaryazdi, Goonesekera and Saffaryazdi, Hailemariam and Temesgen, Nanayakkara and Broadbent and Billinghurst},
   year={2022},
   doi={10.1145/3490099.3511148},
   publisher={27th International Conference on Intelligent User Interfaces (IUI '22), March 22--25, 2022, Helsinki, Finland},
