+++ paper/paper.bib
@@ -121,7 +121,7 @@
 
 @article{Saffaryazdi2022,
   title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
-  author={Saffaryazdi,Nastara and Wasim, Syed Talal and Dileep, Kuldeep and Farrokhinia, Alireza and Nanayakkara, Suranga and Broadbent, Elizabeth, and Billinghurst, Mark},
+  author={Saffaryazdi, Nastaran and Wasim, Syed Talal and Dileep, Kuldeep and Farrokhinia, Alireza and Nanayakkara, Suranga and Broadbent, Elizabeth, and Billinghurst, Mark},
   year={2022},
   deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
   publisher={Manuscript is under review},
