+++ paper/paper.bib
@@ -121,7 +121,7 @@
 
 @article{Saffaryazdi2022,
   title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
-  author={Saffaryazdi, Nastaran and Wasim, Syed Talal and Dileep, Kuldeep and Farrokhinia, Alireza and Nanayakkara, Suranga and Broadbent, Elizabeth and Billinghurst, Mark},
+  author={{Saffaryazdi, Nastaran} and Wasim, Syed Talal and Dileep, Kuldeep and Farrokhinia, Alireza and Nanayakkara, Suranga and Broadbent, Elizabeth and Billinghurst, Mark},
   year={2022},
   deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
   publisher={Manuscript is under review},
@@ -130,7 +130,7 @@
 
 @article{Saffaryazdi2022conv,
   title={Emotion Recognition in Conversations Using Brain and Physiological Signals},
-  author={Saffaryazdi, Nastaran and Goonesekera, Yenushka and Saffaryazdi, Nafiseh and Hailemariam, Nebiyou Daniel and Temesgen, Ebasa Girma and Nanayakkara, Suranga and Broadbent, Elizabeth and Billinghurst, Mark},
+  author={{Saffaryazdi, Nastaran} and Goonesekera, Yenushka and Saffaryazdi, Nafiseh and Hailemariam, Nebiyou Daniel and Temesgen, Ebasa Girma and Nanayakkara, Suranga and Broadbent, Elizabeth and Billinghurst, Mark},
   year={2022},
   doi={10.1145/3490099.3511148},
   publisher={27th International Conference on Intelligent User Interfaces (IUI '22), March 22--25, 2022, Helsinki, Finland},
