+++ paper/paper.bib
@@ -122,6 +122,7 @@
 @article{Saffaryazdi2022,
   title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
   author={Saffaryazdi, Nastaran and Wasim, Syed Talal and Dileep, Kuldeep and Farrokhinia, Alireza and Nanayakkara, Suranga and Broadbent, Elizabeth and Billinghurst, Mark},
+  shortauthor={Saffaryazdi, Nastaran},
   year={2022},
   deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
   publisher={Manuscript is under review},
@@ -131,6 +132,7 @@
 @article{Saffaryazdi2022conv,
   title={Emotion Recognition in Conversations Using Brain and Physiological Signals},
   author={Saffaryazdi, Nastaran and Goonesekera, Yenushka and Saffaryazdi, Nafiseh and Hailemariam, Nebiyou Daniel and Temesgen, Ebasa Girma and Nanayakkara, Suranga and Broadbent, Elizabeth and Billinghurst, Mark},
+  shortauthor={Saffaryazdi, Nastaran},
   year={2022},
   doi={10.1145/3490099.3511148},
   publisher={27th International Conference on Intelligent User Interfaces (IUI '22), March 22--25, 2022, Helsinki, Finland},
@@ -202,6 +204,16 @@
   year={2019},
   publisher={IEEE}
 }
+@article{jiang2020snapshot,
+  title={A snapshot research and implementation of multimodal information fusion for data-driven emotion recognition},
+  author={Jiang, Yingying and Li, Wei and Hossain, M Shamim and Chen, Min and Alelaiwi, Abdulhameed and Al-Hammadi, Muneer},
+  journal={Information Fusion},
+  volume={53},
+  pages={209--221},
+  year={2020},
+  publisher={Elsevier}
+}
+
 
 
 
