+++ paper/paper.bib
@@ -119,25 +119,26 @@
   publisher={IEEE}
 }
 
-@article{Saffaryazdi2022,
-  title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
-  author={Saffaryazdi, Nastaran and Wasim, Syed Talal and Dileep, Kuldeep and Farrokhinia, Alireza and Nanayakkara, Suranga and Broadbent, Elizabeth and Billinghurst, Mark},
-  shortauthor={Saffaryazdi, Nastaran},
-  year={2022},
-  deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
-  publisher={Manuscript is under review},
-  pubstate = {Under review}
-}
-
 @article{Saffaryazdi2022conv,
   title={Emotion Recognition in Conversations Using Brain and Physiological Signals},
   author={Saffaryazdi, Nastaran and Goonesekera, Yenushka and Saffaryazdi, Nafiseh and Hailemariam, Nebiyou Daniel and Temesgen, Ebasa Girma and Nanayakkara, Suranga and Broadbent, Elizabeth and Billinghurst, Mark},
   shortauthor={Saffaryazdi, Nastaran},
-  year={2022},
   doi={10.1145/3490099.3511148},
   publisher={27th International Conference on Intelligent User Interfaces (IUI '22), March 22--25, 2022, Helsinki, Finland},
-  pubstate = {Under publication}
+    year={2022},
+  pubstate={Under publication}
 }
+
+@article{Saffaryazdi2022emotion,
+  title={Using facial micro-expressions in combination with EEG and physiological signals for emotion recognition},
+  author={Saffaryazdi, Nastaran and Wasim, Syed Talal and Dileep, Kuldeep and Farrokhinia, Alireza and Nanayakkara, Suranga and Broadbent, Elizabeth and Billinghurst, Mark},
+  shortauthor={Saffaryazdi, Nastaran},
+  deartment={Empathic Computing Laboratory (ECL), Auckland Bioengineering Intitute, University of Auckland},
+  publisher={Manuscript is under review},
+  year={2022},
+  pubstate={Under review}
+}
+
 
 @article{egger2019emotion,
   title={Emotion recognition from physiological signal analysis: A review},
@@ -210,6 +211,7 @@
   journal={Information Fusion},
   volume={53},
   pages={209--221},
+  doi={10.1016/j.inffus.2019.06.019},
   year={2020},
   publisher={Elsevier}
 }
