+++ paper/paper.md
@@ -65,7 +65,7 @@ The main features of `Octopus Sensing` are listed as follows:
 -![Ovrall view of Octopus Sensing](OCS-diagram.png)
 
 # Research perspective
-We used `Octopus Sensing` to design several human emotion recognition user studies. We developed a user study using `Octopus Sensing` for recording facial video, brain activity, and physiological signals during a watching video task. The recorded data was used to make multimodal emotion recognition models [@Saffaryazdi2022]. This scenario which is common in physiological emotion recognition studies has been included in the repository as an example and explained in the tutorial. In another research, we collected multimodal data during face-to-face conversations to make models for emotion recognition during interactive tasks [@Saffaryazdi2022conv]. We developed this user study's scenario in Python using the `Octopus Sensing` library and conducted the study only by running our developed program, without running any other software or any supervision for data recording or data synchronization.
+We used `Octopus Sensing` to design several human emotion recognition user studies. We developed a user study using `Octopus Sensing` for recording facial video, brain activity, and physiological signals during a watching video task. The recorded data was used to make multimodal emotion recognition models [@Saffaryazdi2022emotion]. This scenario which is common in physiological emotion recognition studies has been included in the repository as an example and explained in the tutorial. In another research, we collected multimodal data during face-to-face conversations to make models for emotion recognition during interactive tasks [@Saffaryazdi2022conv]. We developed this user study's scenario in Python using the `Octopus Sensing` library and conducted the study only by running our developed program, without running any other software or any supervision for data recording or data synchronization.
 
 This tool can be used to build real-time data processing systems to recognize emotions, stress, cognitive load, or analyze human behavior. Our final goal is to extend its capabilities to provide real-time emotion recognition using multimodal data. Furthermore, we plan to integrate it with `Psychopy` in the future and combine multimodal data collection and monitoring with `Psychopy` features when designing scenarios. Additionally, we plan to support LSL in the future. By supporting LSL, other applications that already support LSL could work with `Octopus Sensing`.
 
