### 2022-02-17 8a037ec
Adding image folder for endpoint example
