### 2022-02-17 #4045 @nastaran62
@peircej Thank you very much for your review and comments. We applied the changes as follows.

Community guidelines: We added it to the documents and `.github` repo, so it's available to all three repos.

State of the Field: We updated the paper and also added two issues related to this.

@ixjlyons Thank you very much for your review, comments, and suggestions. We applied the changes as follows.

Documentation
1- Fixed the typos and broken links.
2- Fixed the init script
3- Added troubleshooting for PyGObject
4- Added stimuli docs

Paper, JOSS checklist
1- The third author (Mark Billinghurst) is the primary supervisor of my research, who supports the overall research. He gave feedback about the software and writing. I opened an issue in JOSS and asked how to add him, and the reply was to include him in the authors' list.

2- Rewrote some parts of the paper to fix the grammatical issues and ambiguity.
3- Added the methodology section and a figure
4- Adde community guideline on the documentation website and Github

Other suggestions

1- Great suggestion! Thank you. We are thinking about that and will try to find a solution for this.
2- Included the imports
3- Changed the first example to only use camera instead of Shimmer. For the other ones, since we explained something related to signals, we couldn't change it. For one of the examples, we included both camera and Shimmer in the example file. Users can comment out the not-connected sensor.
4- Added
5- We added a diagram from the main components of the Octopus Sensing to the index page. In the future, we will add more diagrams
6- We added the link to the API

Also, I had an issue with the references that I couldn't resolve. (Sorry for fifty something re-build of the paper above!) When it refer to the references, instead of `Saffaryazdi, et al` it generates `Saffaryazdi, Wasim, et al` and `Saffaryazdi, Goonesekera, et al`. I would appreciate it if someone could have a look.
