### 2022-02-17 #4045 @ajstewartlang
Thanks @peircej and @ixjlyons for those helpful comments - and thanks @nastaran62 for your responses.

:wave: @openjournals/dev Might someone be able to advise re. the referencing glitch in the built PDF please? The expected outcome when the PDF is generated is the following:

`The recorded data was used to make multimodal emotion recognition models (Saffaryazdi, et al., 2022).`

But what is actually generated is the following:

`The recorded data was used to make multimodal emotion recognition models (Saffaryazdi, Wasim, et al., 2022).`

Same issue with the `Saffaryazdi et al.` reference
