### 2022-03-04 #4045 @peircej
Yes, there is now a brief State of the Field section now and all my items are ticked.

@nastaran62 I'm slightly embarassed pointing this out, but there is actually a journal article for PsychoPy that you can cite. It really helps other academic FOSS authors to have citations made properly
https://link.springer.com/article/10.3758/s13428-018-01193-y

Surprisingly, I couldn't find a way to cite Lab Streaming Layer though
