+++ paper/paper.bib
@@ -214,6 +214,16 @@
   publisher={Elsevier}
 }
 
-
+@article{peirce2019psychopy2,
+  title={PsychoPy2: Experiments in behavior made easy},
+  author={Peirce, Jonathan and Gray, Jeremy R and Simpson, Sol and MacAskill, Michael and H{\"o}chenberger, Richard and Sogo, Hiroyuki and Kastman, Erik and Lindel{\o}v, Jonas Kristoffer},
+  journal={Behavior research methods},
+  volume={51},
+  number={1},
+  pages={195--203},
+  year={2019},
+  doi={10.3758/s13428-018-01193-y},
+  publisher={Springer}
+}
 
 
