### 2022-03-04 #4045 @nastaran62
@peircej Thank you very much for your review. Sorry about my mistake. I cited the PsychoPy paper too. I also couldn't find any original paper for LSL to cite. There are only some papers that used LSL.
