### 2022-03-05 #4045 @ixjlyons
I have just one minor comment on the most recent paper version: line 92 (first paragraph of Research perspective section), maybe change "In another research" to "In another study" or "In another research study".

Aside from that, everything looks good to me. Nice job on the figure.
