### 2022-03-07 #4045 @nastaran62
@ajstewartlang, Yes I did. Sorry, I forgot to regenerate the PDF. I could see the correct one in the pdf that you generated and it seems correct.
About the invalid DOI, it is related to my paper that would be published by the end of this month, and still, its DOI is inactive. In the following issue, I asked how to cite it and Afron answered "Feel free to cite with the DOI that is not yet available". I'm sorry I should have mentioned it earlier.

https://github.com/openjournals/joss/issues/1042#issuecomment-1041878032

Thank you very much.
