### 2022-03-08 #4045 @ajstewartlang
:wave: @nastaran62 I'm just having a final pass through and have spotted a few small things in the reference list that need correcting.

- For the Chen et al. (2020) paper could you add the volume/issue/page numbers please after the journal title to be consistent with the rest of the references: Journal of Personality and Social Psychology, 121(5), 1029–1056
- Looks like there is a formatting typo in the Hong et al. (2020) reference - "bidirectional communication\vspace* 7pt"
- For Saffaryazdi, N., Goonesekera, Y. et al. (2022) could you change 'Under publication' to 'In press'
- Ensure all references follow the same format and include page numbers where applicable - for example, I spotted page numbers available but missing for the Mangaroska et al. (2021), and the Verschuere et al. (2006) references

Thanks!
