+++ paper/paper.bib
@@ -1,8 +1,8 @@
 @article{kreibig2010autonomic,
   title={Autonomic nervous system activity in emotion: A review},
   author={Kreibig, Sylvia D},
-  journal={Biological psychology},
   doi={10.1016/j.biopsycho.2010.03.010},
+  journal={Biological psychology},
   volume={84},
   number={3},
   pages={394--421},
@@ -10,20 +10,23 @@
   publisher={Elsevier}
 }
 
-@article{chen2020physiological,
+@article{chen2021physiological,
   title={Physiological linkage during shared positive and shared negative emotion.},
   author={Chen, Kuan-Hua and Brown, Casey L and Wells, Jenna L and Rothwell, Emily S and Otero, Marcela C and Levenson, Robert W and Fredrickson, Barbara L},
   doi={10.1037/pspi0000337},
   journal={Journal of Personality and Social Psychology},
-  year={2020},
+  volume={121},
+  number={5},
+  pages={1029},
+  year={2021},
   publisher={American Psychological Association}
 }
 
 @article{sun2020multimodal,
   title={Multimodal Affective State Assessment Using fNIRS+ EEG and Spontaneous Facial Expression},
   author={Sun, Yanjia and Ayaz, Hasan and Akansu, Ali N},
-  journal={Brain Sciences},
   doi={10.3390/brainsci10020085},
+  journal={Brain Sciences},
   volume={10},
   number={2},
   pages={85},
@@ -34,8 +37,8 @@
 @article{hassouneh2020development,
   title={Development of a Real-Time Emotion Recognition System Using Facial Expressions and EEG based on machine learning and deep neural network methods},
   author={Hassouneh, Aya and Mutawa, AM and Murugappan, M},
-  journal={Informatics in Medicine Unlocked},
   doi={10.1016/j.imu.2020.100372},
+  journal={Informatics in Medicine Unlocked},
   pages={100372},
   year={2020},
   publisher={Elsevier}
@@ -44,8 +47,8 @@
 @article{verschuere2006psychopathy,
   title={Psychopathy and physiological detection of concealed information: A review},
   author={Verschuere, Bruno and Crombez, Geert and Koster, Ernst and Uzieblo, Katarzyna},
-  journal={Psychologica Belgica},
   doi={10.5334/pb-46-1-2-99},
+  journal={Psychologica Belgica},
   volume={46},
   number={1-2},
   year={2006},
@@ -55,8 +58,8 @@
 @article{hossain2019observers,
   title={Observers’ physiological measures in response to videos can be used to detect genuine smiles},
   author={Hossain, Md Zakir and Gedeon, Tom},
-  journal={International Journal of Human-Computer Studies},
   doi={10.1016/j.ijhcs.2018.10.003},
+  journal={International Journal of Human-Computer Studies},
   volume={122},
   pages={232--241},
   year={2019},
@@ -66,8 +69,8 @@
 @article{dewan2019engagement,
   title={Engagement detection in online learning: a review},
   author={Dewan, M Ali Akber and Murshed, Mahbub and Lin, Fuhua},
-  journal={Smart Learning Environments},
   doi={10.1186/s40561-018-0080-z},
+  journal={Smart Learning Environments},
   volume={6},
   number={1},
   pages={1},
@@ -78,28 +81,34 @@
 @article{val2020affective,
   title={Affective Robot Story-Telling Human-Robot Interaction: Exploratory Real-Time Emotion Estimation Analysis Using Facial Expressions and Physiological Signals},
[…]
