+++ paper/paper.bib
@@ -158,13 +158,15 @@
   publisher={Elsevier}
 }
 
-@article{vanneste2020towards,
+@article{vanneste2021towards,
   title={Towards measuring cognitive load through multimodal physiological data},
   author={Vanneste, Pieter and Raes, Annelies and Morton, Jessica and Bombeke, Klaas and Van Acker, Bram B and Larmuseau, Charlotte and Depaepe, Fien and Van den Noortgate, Wim},
   doi={10.1007/s10111-020-00641-0},
   journal={Cognition, Technology \& Work},
-  pages={1--19},
-  year={2020},
+  volume={23},
+  number={3},
+  pages={567--585},
+  year={2021},
   publisher={Springer}
 }
 
