+++ paper/paper.bib
@@ -51,6 +51,7 @@
   journal={Psychologica Belgica},
   volume={46},
   number={1-2},
+  pages={99},
   year={2006},
   publisher={Ubiquity Press}
 }
