### 2022-03-09 #4045 @ajstewartlang
Many thanks @nastaran62 - the paper now looks good to me. Thanks for submitting this great work to JOSS - and many thanks again to @peircej and @ixjlyons for reviewing!
