### 2022-03-09 #4045 @ajstewartlang
:wave: @nastaran62 If you could now do the following please, that would be great:

- Make a [tagged release](https://docs.github.com/en/github/administering-a-repository/releasing-projects-on-github/managing-releases-in-a-repository) of the software.
