### 2022-03-09 48247c9
add test_brainflow_streaming with csv-save error
