### 2022-03-13 01c6f61
Fix some comments and prints
