+++ paper/paper.bib
@@ -39,6 +39,7 @@
   author={Hassouneh, Aya and Mutawa, AM and Murugappan, M},
   doi={10.1016/j.imu.2020.100372},
   journal={Informatics in Medicine Unlocked},
+  volume={20},
   pages={100372},
   year={2020},
   publisher={Elsevier}
