+++ octopus_sensing/tests/test_brainflow_streaming.py
@@ -1,9 +1,8 @@
 import pytest
 import multiprocessing
-import multiprocessing.dummy
 import multiprocessing.queues
+import threading
 import queue
-import random
 import http.client
 import pickle
 import os
@@ -12,6 +11,12 @@ import numpy as np
 import time
 import tempfile
 from brainflow import board_shim
+
+import octopus_sensing.devices.brainflow_streaming as brainflow_streaming
+from octopus_sensing.device_coordinator import DeviceCoordinator
+from octopus_sensing.common.message_creators import start_message, stop_message, terminate_message
+from octopus_sensing.monitoring_endpoint import MonitoringEndpoint
+
 
 class MockBrainFlowInputParams():
     def __init__(self):
@@ -44,80 +49,65 @@ class MockBoardShim():
 
 @pytest.fixture(scope="module")
 def mocked():
-    # Preventing processes from creating a new process
-    original_process = multiprocessing.Process
-    multiprocessing.Process = multiprocessing.dummy.Process
-    original_queue = multiprocessing.queues.Queue
-    multiprocessing.queues.Queue = queue.Queue
-    original_queue_method = multiprocessing.Queue
-    multiprocessing.Queue = queue.Queue
-
-    original_BrainFlowInputParams = board_shim.BrainFlowInputParams
-    board_shim.BrainFlowInputParams = MockBrainFlowInputParams
-    original_BoardShim = board_shim.BoardShim
-    board_shim.BoardShim = MockBoardShim
+    original_bases = brainflow_streaming.BrainFlowStreaming.__bases__[0].__bases__[0].__bases__
+    brainflow_streaming.BrainFlowStreaming.__bases__[0].__bases__[0].__bases__ = (threading.Thread,)
+    original_BrainFlowInputParams = brainflow_streaming.BrainFlowInputParams
+    brainflow_streaming.BrainFlowInputParams = MockBrainFlowInputParams
+    original_BoardShim = brainflow_streaming.BoardShim
+    brainflow_streaming.BoardShim = MockBoardShim
 
     yield None
 
-    multiprocessing.Process = original_process
-    multiprocessing.queues.Queue = original_queue
-    multiprocessing.Queue = original_queue_method
-    board_shim.BrainFlowInputParams = original_BrainFlowInputParams
-    board_shim.BoardShim = original_BoardShim
+    brainflow_streaming.BrainFlowStreaming.__bases__[0].__bases__[0].__bases__ = original_bases
+    brainflow_streaming.BrainFlowInputParams = original_BrainFlowInputParams
+    brainflow_streaming.BoardShim = original_BoardShim
 
 def test_system_health(mocked):
-    import octopus_sensing.devices.brainflow_streaming as brainflow_streaming
-    from octopus_sensing.device_coordinator import DeviceCoordinator
-    from octopus_sensing.common.message_creators import start_message, stop_message, terminate_message
-    from octopus_sensing.monitoring_endpoint import MonitoringEndpoint
 
     output_dir = tempfile.mkdtemp(prefix="octopus-sensing-test")
-    print(output_dir)
-    coordinator = DeviceCoordinator()
+    experiment_id = 'test-exp-2'
+    stimuli_id = 'sti-2'
 
     params = board_shim.BrainFlowInputParams()
     params.serial_port = "/dev/ttyUSB0"
-    brainflow = \
+    device = \
         brainflow_streaming.BrainFlowStreaming(2,
                                                125,
                                                brain_flow_input_params=params,
                                                name="cyton_daisy",
                                                output_path=output_dir)    
+    msg_queue = queue.Queue()
+    device.set_queue(msg_queue)
+    monitoring_queue_in = queue.Queue()
+    monitoring_queue_out = queue.Queue()
[…]
