+++ octopus_sensing/__init__.py
@@ -12,4 +12,4 @@
 # You should have received a copy of the GNU General Public License along with Octopus Sensing.
 # If not, see <https://www.gnu.org/licenses/>.
 
-__version__ = '3.1.0'
+__version__ = '3.1.1'
