### 2022-03-13 90bc001
Github actions: Run on Pull Request too
