+++ .github/workflows/python-check.yml
@@ -4,7 +4,7 @@
 
 name: Python Check
 
-on: [push]
+on: [push, pull_request]
 
 jobs:
   build:
