### 2022-03-14 #4045 @ajstewartlang
Just to highlight @openjournals/dev the invalid DOI warning was discussed [here](https://github.com/openjournals/joss/issues/1042#issuecomment-1041878032) with @arfon giving the thumbs up to citing without that DOI being available.
