### 2022-03-23 d7d2521
Update paper.md
Fix a typo
